Server-AI/requirements.txt
numpy==1.20.1
meteostat==1.1.1
pandas==1.2.2
parquet==1.3.1
Pillow==8.1.1
pyproj==3.0.0.post1
scikit_learn==0.24.1
tensorflow==2.4.1
